app/pyproject.toml
[project]
name = "text2emoji"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "google-genai",
    "python-dotenv",
    "pydantic"
]
